int -> size_t

+++ tools/extract_proto/extract_proto.cc
@@ -285,8 +285,8 @@ int PopulatePartitionTree(insp_frame_data *frame_data, insp_sb_data *sb_data,
   int coding_unit_end = coding_unit_start;
 
   int child_coding_unit_start = coding_unit_start;
-  int i = 0;
-  for (; i < (int)std::size(tree->sub_tree); ++i) {
+  size_t i = 0;
+  for (; i < std::size(tree->sub_tree); ++i) {
     if (tree->sub_tree[i] == nullptr) {
       break;
     }
